+++ src/exec.c
@@ -694,7 +694,7 @@ nxtscan:	/* on to the next line */
 		/* if macro store is on, just salt this away */
 		if (mstore) {
 			/* allocate the space for the line */
-			linlen = strlen(eline);
+			linlen = (int)strlen(eline);
 			if ((mp=lalloc(linlen)) == NULL) {
 				errormesg(TEXT125, bp, lp);
 /*                                      "Out of memory while storing macro" */
@@ -796,7 +796,7 @@ nxtscan:	/* on to the next line */
 #endif
 					/* grab label to jump to */
 					eline = token(eline, golabel, NPAT);
-					linlen = strlen(golabel);
+					linlen = (int)strlen(golabel);
 					glp = lforw(hlp);
 					while (glp != hlp) {
 						if ((lused(glp) >= linlen) &&
